xfree86: Allow sdksyms.dep to be included portably
Non-GNU makes don't deal with the sinclude or -include variants that
allow Makefile stubs to be created and then included during the build.
Instead, create an empty file at the end of configure so that the
regular include statement can be included. This is how automake handles
automatic source dependencies.

In order to trick automake into not processing the include statement, a
variable is used.

Signed-off-by: Dan Nicholson <dbn.lists@gmail.com>
Tested-by: Alan Coopersmith <alan.coopersmith@oracle.com>