hw/xwin: Look for gl spec files in $(srcdir) or .
Tarballs include the downloaded gl spec files, which will end up in
$(srcdir). But, git-based builds will not have them at all and will
need to download them from opengl.org. They'll land in in the build
directory instead of $(srcdir), and so we need to allow them to be in
either place.

This change checks for the files in $(srcdir), linking them to . if
present. Otherwise, it downloads them from opengl.org.

A suggested better solution is to have Mesa install these files somewhere.

Signed-off-by: Keith Packard <keithp@keithp.com>
Tested-by: Kristian Høgsberg <krh@bitplanet.net>