hw/dmx/doc: Add explicit dependency for all doxygen output files
Instead of listing one of the doxygen output files and depending on
sequential execution to ensure that the other files were present
before make checked for them, create explicit dependencies so that
make will not check for the additional files until after doxygen has
been run.

This allows parallel make to work correctly in this directory.

Signed-off-by: Keith Packard <keithp@keithp.com>
Tested-by: Kristian Høgsberg <krh@bitplanet.net>