xfixes: Mark some variables as unused.
Calling Unwrap() is just a way of performing an assignment while keeping a
backup of the original value. In the CursorCloseScreen function, the backup
value doesn't seem to be useful, but Unwrap() is used anyway (probably to stay
in line with other functions). As a consequence, mark those variables as unused.

The following warnings go away accordingly:
|   CC     cursor.lo
| cursor.c: In function 'CursorCloseScreen':
| cursor.c:186:26: warning: variable 'display_proc' set but not used [-Wunused-but-set-variable]
| cursor.c:185:24: warning: variable 'close_proc' set but not used [-Wunused-but-set-variable]

Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>
Signed-off-by: Cyril Brulebois <kibi@debian.org>