glx: Add _glapi_create_table_from_handle
_glapi_create_table_from_handle was recently added to mesa to ease creation of
a _glapi_table.  This commit brings this API over to xserver for use in
XQuartz (next commit).

The API was generated with scripts from current mesa followed by various
script-foo to remove entries that are not in the server and add those which
are still in the server but not in mesa (CullParameterdvEXT and
CullParameterfvEXT)

Signed-off-by: Jeremy Huddleston <jeremyhu@apple.com>
Reviewed-by: Adam Jackson <ajax@redhat.com>