glamor: Fixed one bug when enable dynamic pixmap uploading.
When try to upload a pixmap without yInverted set, we must
set up a fbo for it to do the y flip. Previous implementation
only consider the ax bit. After fix this problem, we can
enable the dynamic uploading feature in copyarea function when
the yInverted is not set (from Xephyr).

Signed-off-by: Zhigang Gong <zhigang.gong@gmail.com>